+++ tests/e2e/specs/test.js
@@ -8,6 +8,10 @@ module.exports = {
       .waitForElementVisible('#app', 5000)
       .assert.elementPresent('.navbar')
       .assert.elementPresent('.navbar-brand')
+      .assert.elementPresent('.navbar-expand-md')
+      .resizeWindow(200, 400)
+      .waitForElementVisible('.navbar-toggler-icon', 5000)
+      .assert.cssClassNotPresent('.navbar.navbar-light.bg-light', '.navbar-expand-md')
       .end()
   },
   'display cookiewall': browser => {