+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Dynamic (via js) menu collapse breakpoints (hamburger menu).
 ### Removed
 ### Changed
+- Larger logo height, set via css on HTML class
 ### Fixed
 
 ## [0.2.0] - 2019-01-11